LPS-55754 move the rest to match the pattern in Journal
