+++ modules/apps/wiki/wiki-web/bnd.bnd
@@ -39,7 +39,6 @@ Import-Package:\
 	com.liferay.wiki.configuration,\
 	com.liferay.wiki.constants,\
 	com.liferay.wiki.importer.impl,\
-	com.liferay.wiki.indexer,\
 	com.liferay.wiki.lar,\
 	com.liferay.wiki.model.impl,\
 	com.liferay.wiki.notifications,\
